Restart get processes if they don't get enough responses
With the recent changes to reduce load on the bootstrap servers it has
become more likely that a process will get a set of stale nodes at
startup. To mitigate this allow processes to pull a new set of nodes
from the routing table and restart if they don't get at least a bucket's
worth of responses.

Review: http://cr.bittorrent.com/5750322559451136/